% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allbus2012}
\alias{allbus2012}
\title{allbus2012}
\format{
A data frame with 3480 rows and 3 variables:
\enumerate{
\item \strong{v220} age of the respondent (years)
\item \strong{v593} body height of the respondent (cm)
\item \strong{v595} weight of the respondent (kg)
}
}
\usage{
allbus2012
}
\description{
An exercept of three variables of the ALLBUS 2012 data set.
}
\references{
GESIS - Leibniz-Institut für Sozialwissenschaften (2013). Allgemeine Bevölkerungsumfrage der Sozialwissenschaften ALLBUS 2012. GESIS Datenarchiv, Köln. ZA4614 Datenfile Version 1.1.1, https://doi.org/10.4232/1.11753.
}
\keyword{datasets}
