% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allbus2018}
\alias{allbus2018}
\title{allbus2018}
\format{
A data frame with 3477 rows and 20 variables:
\enumerate{
\item \strong{eastwest} SURVEY AREA (RESIDENCE AREA): WEST - EAST
\item \strong{ep03} ECONOMIC SITUATION OF RESPONDENT TODAY
\item \strong{ep06} ECONOMIC SITUATION OF RESPONDENT IN 1 YEAR
\item \strong{di05} HOUSEHOLD NET INCOME: OPEN QUESTION
\item \strong{di06} HOUSEHOLD NET INCOME: LIST QUESTION
\item \strong{hhinc} HOUSEHOLD NET INCOME (OPEN+LISTS INPUT), CATEGORIES
\item \strong{land} FEDERAL STATE WHERE RESPONDENT RESIDES
\item \strong{pt01} TRUST: HEALTH CARE
\item \strong{pt02} TRUST: FEDERAL CONSTITUTIONAL COURT
\item \strong{pt03} TRUST: FEDERAL PARLIAMENT
\item \strong{pt04} TRUST: CITY, MUNICIPAL ADMINISTRATION
\item \strong{pt08} TRUST: JUSTICE
\item \strong{pt09} TRUST: TELEVISION
\item \strong{pt10} TRUST: NEWSPAPER
\item \strong{pt11} TRUST: HIGHER EDUCATION, UNIVERSITIES
\item \strong{pt12} TRUST: FEDERAL GOVERNMENT
\item \strong{pt14} TRUST: POLICE
\item \strong{pt15} TRUST: POLITICAL PARTIES
\item \strong{pt19} TRUST: COMMISSION OF THE EU
\item \strong{pt20} TRUST: EUROPEAN PARLIAMENT
}
}
\usage{
allbus2018
}
\description{
An exercept of some variables of the ALLBUS 2018 data set.
}
\references{
GESIS - Leibniz-Institut für Sozialwissenschaften (2019). German General Social Survey - ALLBUS 2018. GESIS Datenarchiv, Köln. ZA5272 Datenfile Version 1.0.0, https://doi.org/10.4232/1.13325.
}
\keyword{datasets}
