% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anscombe}
\alias{anscombe}
\title{anscombe}
\format{
A data frame with 3477 rows and 20 variables:
\enumerate{
\item \strong{ID} number of data set
\item \strong{X} x
\item \strong{Y} y
}
}
\source{
\link[datasets:anscombe]{datasets::anscombe}
}
\usage{
anscombe
}
\description{
Four artificial data sets which have the same traditional statistical properties
(mean, variance, correlation, regression line, etc.), yet are quite different.
}
\references{
Tufte, Edward R. (1989). The Visual Display of Quantitative Information, 13–14. Graphics Press.
}
\keyword{datasets}
