% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/askUser.R
\name{askUser}
\alias{askUser}
\title{askUser}
\usage{
askUser(msg, choices = c("Yes", "No", "Cancel"), default = 1)
}
\arguments{
\item{msg}{character: the prompt message for the user}

\item{choices}{character: vector of choices (default: \code{c("Yes", "No", "Cancel")})}

\item{default}{character/integer: default option if only \code{Enter} pressed (default: \code{1})}
}
\value{
the integer number choosen by the user
}
\description{
\code{askUser} provides a way to ask the user a yes/no/cancel question (default). A \code{*} after a number indicates the default option.
}
\examples{
if (interactive())
  askUser("Do you want to use askUser?")
}
