% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bank2}
\alias{bank2}
\title{bank2}
\format{
A data frame with 200 rows and 6 variables:
\enumerate{
\item \strong{WIDTH} width of bill (mm)
\item \strong{LEFT} width of left edge (mm)
\item \strong{RIGHT} width of right edge (mm)
\item \strong{UPPER} upper margin width (mm)
\item \strong{LOWER} lower margin width (mm)
\item \strong{DIAGONAL} length of diagonal (mm)
}
}
\source{
Flury, B. and Riedwyl, H. (1988). Multivariate Statistics: A practical approach. London: Chapman & Hall, Tables 1.1 and 1.2, pp. 5-8.
}
\usage{
bank2
}
\description{
The data set contains six measurements made on 100 genuine and 100 counterfeit old-Swiss 1000-franc bank notes.
}
\keyword{datasets}
