% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benzinverbrauch}
\alias{benzinverbrauch}
\title{benzinverbrauch}
\format{
A data frame with 200 rows and 6 variables:
\enumerate{
\item \strong{Klasse} fuel consumption (liter/100 km)
\item \strong{h(x)} absolute frequency
\item \strong{f(x)} relative frequency
}
}
\usage{
benzinverbrauch
}
\description{
The data set with grouped data of fuel consumption of cars.
}
\keyword{datasets}
