% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bostonc}
\alias{bostonc}
\title{bostonc}
\format{
A data frame with 506 rows and 20 variables:
\enumerate{
\item \strong{TOWN} name of town
\item \strong{TOWN#} number of town
\item \strong{TRACT} census tract
\item \strong{LON}	longitude of census tract
\item \strong{LAT} latitude of census tract
\item \strong{MEDV} median value of owner-occupied homes in USD 1000's
\item \strong{CMEDV} corrected median value of owner-occupied homes in USD 1000's
\item \strong{CRIM}	per capita crime rate by town
\item \strong{ZN} proportion of residential land zoned for lots over 25,000 sq.ft
\item \strong{INDUS} proportion of non-retail business acres per town
\item \strong{CHAS}	Charles River dummy variable (= 1 if tract bounds river; 0 otherwise)
\item \strong{NOX} nitric oxides concentration (parts per 10 million)
\item \strong{RM}	average number of rooms per dwelling
\item \strong{AGE} proportion of owner-occupied units built prior to 1940
\item \strong{DIS} weighted distances to five Boston employment centres
\item \strong{RAD} index of accessibility to radial highways
\item \strong{TAX} full-value property-tax rate per USD 10,000
\item \strong{PTRATIO} pupil-teacher ratio by town
\item \strong{B} 1000(B - 0.63)^2 where B is the proportion of blacks by town
\item \strong{LSTAT} percentage of lower status of the population
}
}
\source{
\href{http://lib.stat.cmu.edu/datasets/boston_corrected.txt}{From Statlib: boston_corrected}. Submitted by Kelley Pace (kpace@unix1.sncc.lsu.edu).
}
\usage{
bostonc
}
\description{
Housing data for 506 census tracts of Boston from the 1970 census. It contains the corrected version with additional spatial information.
}
\references{
\itemize{
\item Harrison, David, and Daniel L. Rubinfeld, Hedonic Housing Prices and the Demand for Clean Air, Journal of Environmental Economics and Management, Volume 5, (1978), 81-102.
\item Gilley, O.W., and R. Kelley Pace, On the Harrison and Rubinfeld Data, Journal of Environmental Economics and Management, 31 (1996), 403-405. Provided corrections and examined censoring.
\item Pace, R. Kelley, and O.W. Gilley, Using the Spatial Configuration of the Data to Improve Estimation, Journal of the Real Estate Finance and Economics 14 (1997), 333-340. Added georeferencing and spatial estimation.
}
}
\seealso{
\link[mlbench:BostonHousing]{mlbench::BostonHousing}, \link[spData:boston]{spData::boston}
}
\keyword{datasets}
