% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bostonh}
\alias{bostonh}
\title{bostonh}
\format{
A data frame with 506 rows and 20 variables:
\enumerate{
\item \strong{CRIM}	per capita crime rate by town
\item \strong{ZN} proportion of residential land zoned for lots over 25,000 sq.ft
\item \strong{INDUS} proportion of non-retail business acres per town
\item \strong{CHAS}	Charles River dummy variable (= 1 if tract bounds river; 0 otherwise)
\item \strong{NOX} nitric oxides concentration (parts per 10 million)
\item \strong{RM}	average number of rooms per dwelling
\item \strong{AGE} proportion of owner-occupied units built prior to 1940
\item \strong{DIS} weighted distances to five Boston employment centres
\item \strong{RAD} index of accessibility to radial highways
\item \strong{TAX} full-value property-tax rate per USD 10,000
\item \strong{PTRATIO} pupil-teacher ratio by town
\item \strong{B} 1000(B - 0.63)^2 where B is the proportion of blacks by town
\item \strong{LSTAT} percentage of lower status of the population
\item \strong{MEDV} median value of owner-occupied homes in USD 1000's
}
}
\source{
\href{http://lib.stat.cmu.edu/datasets/boston_corrected.txt}{From Statlib: boston_corrected}. Submitted by Kelley Pace (kpace@unix1.sncc.lsu.edu).
}
\usage{
bostonh
}
\description{
Housing data for 506 census tracts of Boston from the 1970 census. It contains the original data by Harrison and Rubinfeld (1979),
}
\references{
Harrison, David, and Daniel L. Rubinfeld, Hedonic Housing Prices and the Demand for Clean Air, Journal of Environmental Economics and Management, Volume 5, (1978), 81-102.
}
\seealso{
\link[MASS:Boston]{MASS::Boston}, \link[mlbench:BostonHousing]{mlbench::BostonHousing}
}
\keyword{datasets}
