% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{concrete}
\alias{concrete}
\title{concrete}
\format{
A data frame with 103 rows and 10 variables:
\enumerate{
\item \strong{Cement} Cement (in kg/m$^3$ concrete)
\item \strong{Slag} Slag (in kg/m$^3$ concrete)
\item \strong{Flyash} Fly ash (in kg/m$^3$ concrete)
\item \strong{Water} Water (in kg/m$^3$ concrete)\\
\item \strong{SP} Superplasticizer (in kg/m$^3$ concrete)
\item \strong{CoarseAggr} Coarse sand (in kg/m$^3$ concrete)
\item \strong{FineAggr} Fine sand (in kg/m$^3$ concrete)
\item \strong{Slump} Slump (in cm)
\item \strong{Flow} Flow (in cm)
\item \strong{Strength} 28day compressive strength (in Megapascal)
}
}
\source{
\href{https://archive.ics.uci.edu/ml/datasets/concrete+compressive+strength}{UCI Machine Learning Repository - Concrete Compressive Strength Data Set}
}
\usage{
concrete
}
\description{
Data about concrete mixtures and resulting properties (Slump, Flow and Strength).
Concrete is the most important material in civil engineering. The
concrete compressive strength is a highly nonlinear function of age and
water, superplasticizer, coarse aggregate, and fine aggregate.
}
\details{
\itemize{
\item Water content: ACI 211.1 (Standard practice for selecting proportions for normal, heavyweight, and mass concrete) assumes that, for a given maximum size of coarse aggregate, the slump or consistency of concrete is a direct function of the water content; i.e., within limits it is independent of other factors such as aggregate grading and cement content.
\item Superplasticizer:  When the water content of a concrete mixture is held constant, the addition of a superplasticizer will increase the consistency.
\item Pozzolanic admixtures: Pozzolanic admixtures tend to improve the cohesiveness of concrete. Fly ash, when used as a partial replacement for fine aggregate, generally increases the consistency at a given water content.
\item Cement content: Concretes containing a very high proportion of cement show excellent cohesiveness, but tend to be sticky. At a given water content, a considerable lowering of the cement content tends to produce harsh mixtures, with poor cohesiveness.
\item Aggregate characteristics: Very fine sands require more water for a given consistency; alternatively, they will produce harsh and unworkable mixtures at a water content that might have been adequate with coarser sands. Also, the particle size of coarse aggregate influences the water requirement for a given consistency.
}
}
\references{
I-Cheng Yeh (1998), Modeling of strength of high performance concrete using artificial neural networks, Cement and Concrete Research, Vol. 28, No. 12, pp. 1797-1808
}
\keyword{datasets}
