% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country}
\alias{country}
\title{country
The data set contains data about 122 countries in 1992.}
\format{
A data frame with 122 rows and 22 variables:
\enumerate{
\item \strong{COUNTRY} Country name
\item \strong{POP92} Population, 1992, in millions
\item \strong{URBAN} Percent urban, 1992
\item \strong{GDP} GDP per capita
\item \strong{LIFEEXPM} Male life expectancy 1992
\item \strong{LIFEEXPF} Female life expectancy 1992
\item \strong{BIRTHRAT} Births per 1000 population, 1992
\item \strong{DEATHRAT} Deaths per 1000 people, 1992
\item \strong{INFMR} Infant mortality rate 1992 (per 1000 live births)
\item \strong{FERTRATE} Fertility rate per woman, 1990
\item \strong{REGION} Region of the world (1=Eastern Africa, 2=Middle Africa, 3=Northern Africa, 4=Southern Africa, 5=Western Africa, 6=Caribbean, 7=Central America, 8=South America, 9=North America, 10=Eastern Asia, 11=Southeast Asia, 12=Southern Asia, 13=Western Asia, 14=Eastern Europe, 15=Northern Europe, 16=Southern Europe, 17=Western Europe, 18=Oceania, 19=USSR)
\item \strong{DEVELOP} Status as Developing Country (0=Developed country, 1=Developing country)
\item \strong{RADIO} Radios per 100 people
\item \strong{PHONE} Phones per 100 people
\item \strong{HOSPBED} Hospital beds per 10,000 people
\item \strong{DOCS} Doctors per 10,000 people
\item \strong{LNDOCS} Natural log of doctors per 10000
\item \strong{LNRADIO}  Natural log of radios per 100 people
\item \strong{LNPHONE} Natural log of phones per 100 people
\item \strong{LNGDP} Natural log of
\item \strong{SEQUENCE} Arbitrary sequence number
\item \strong{LNBEDS} Natural log hospital beds/10,000
}
}
\source{
\href{http://www.norusis.com/book_DA_v19.php}{Norušis, Marija J. IBM SPSS Statistics 19: Guide to Data Analysis. Internat. ed. Upper Saddle River: Prentice Hall, 2011.}
}
\usage{
country
}
\description{
country
The data set contains data about 122 countries in 1992.
}
\keyword{datasets}
