% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cps78_85}
\alias{cps78_85}
\title{cp75_85}
\format{
A data frame with 1084 rows and 11 variables:
\enumerate{
\item \strong{educ} Number of years of education
\item \strong{south} Indicator variable for Southern Region (1=Person lives in South, 0=Person lives elsewhere)
\item \strong{nonwhite} Race (0=White, 1=Other)
\item \strong{female} Indicator variable for sex (1=Female, 0=Male)
\item \strong{married} Marital Status (0=Unmarried,  1=Married)
\item \strong{exper} Number of years of work experience
\item \strong{expersq} Number of years of work experience squared
\item \strong{union} Indicator variable for union membership (1=Union member, 0=Not union member)
\item \strong{lwage} Logarithm of wage (dollars per hour)
\item \strong{age} Age (years)
\item \strong{year} Survey year (78=1978, 85=1985)
}
}
\source{
\href{http://lib.stat.cmu.edu/datasets/CPS_85_Wages}{Statlib: Determinants of Wages from the 1985 Current Population Survey}, 1978 Current Population Survey (???)
}
\usage{
cps78_85
}
\description{
The Current Population Survey (CPS) is used to supplement census information between census years. These data consist of a random sample of 534 persons from the CPS 1985 and 550 persons from the CPS 1978, with information on wages and other characteristics of the workers, including sex, number of years of education, years of work experience, occupational status, region of residence and union membership. We wish to determine (i) whether wages are related to these characteristics and (ii) whether there is a gender gap in wages.
}
\references{
Berndt, ER. (1991) The Practice of Econometrics, NY: Addison-Wesley.
}
\keyword{datasets}
