% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{credit}
\alias{credit}
\title{credit}
\format{
A data frame with 1000 rows and 8 variables:
\enumerate{
\item \strong{Problems} Credit repaid
\item \strong{BankAccount} Current account with the bank (no/bad running/good running)
\item \strong{Duration} Duration in months
\item \strong{PreviousPayment} Previous payment record (bad customer/good customer)
\item \strong{CreditUse} Intended purpose (private/professional)
\item \strong{DM} Loan amount in DM
\item \strong{Sex} Gender (female/male)
\item \strong{CivilState} Marital status (married/not married)
}
}
\source{
Exercept from \href{https://doi.org/10.5282/ubm/data.23}{Kreditscoring zur Klassifikation von Kreditnehmern. 2010. Open Data LMU.}
}
\usage{
credit
}
\description{
The granting of personal loans is made dependent on the creditworthiness of the customers,
i.e. on the willingness and ability of the customers to duly pay the interest and redemption
instalments due. A potential borrower can be classified either as a problem-free borrower or
as a problematic borrower who should be checked more closely or who should not be granted a loan.
Each borrower is characterised by a series of features that characterise his personal, economic
and legal situation. On the basis of these characteristics, an attempt is made to make a statistically
sound decision on granting or rejecting the loan.
The data are a stratified learning sample conducted by a major bank in southern Germany.
With a population of 1000 former borrowers, mainly ordinal and nominal survey characteristics were recorded
to describe their living conditions. 30 percent of these former customers were not able to repay the
loan as agreed.
}
\references{
\itemize{
\item Fahrmeir, L. / Tutz, G. (1994): Multivariate Statistical Modelling Based on Generalized Linear Models. Springer, New York.
\item Fahrmeir, L. / Hamerle, A. / Tutz, G. (1984, 1. Aufl.): Multivariate statistische Verfahren. de Gruyter, Berlin.
\item Fahrmeir, L. / Hamerle, A. / Tutz, G. (1996, 2. Aufl.): Multivariate statistische Verfahren. de Gruyter, Berlin.
\item Fahrmeir, L. / Hamerle, A. (1981): Kategoriale Regression in der betrieblichen Planung. Zeitschrift für Operations Research, 44/B, 63-78.
\item Feilmeier, M. / Fergel, I. / Segerer, G. (1981): Lineare Diskriminanz- und Clusteranalyseverfahren bei Kreditscoring-Systemen. Zeitschrift für Operations Research, 25/B, 25-28.
}
}
\keyword{datasets}
