% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data12411}
\alias{data12411}
\title{data12411}
\format{
A data frame with 86 rows and 3 variables:
\enumerate{
\item \strong{V1}  age (in years)
\item \strong{M}  total number of males in age group
\item \strong{F}  total number of females in age group
}
}
\source{
\href{https://www-genesis.destatis.de/genesis/online/}{GENESIS-Online Database}
}
\usage{
data12411
}
\description{
Number of males and females by age.
}
\keyword{datasets}
