% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{defaultApp}
\alias{defaultApp}
\title{defaultApp}
\usage{
defaultApp(file)
}
\arguments{
\item{file}{character: file name}
}
\value{
nothing
}
\description{
Tries to open the given \code{file} with the default application of the operating system.
Only Windows (\code{windows}), macOS (\code{darwin}) and Linux (\code{linux}) is supported.
}
\examples{
if (interactive()) {
  ghget()
  defaultApp(ghlist("dataanalysis.pdf", full.names = TRUE))
}
}
