% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{destatis13211_0002}
\alias{destatis13211_0002}
\title{destatis13211_0002}
\format{
A data frame with 86 rows and 3 variables:
\enumerate{
\item \strong{Gebiet}  territory
\item \strong{Jahr}  year
\item \strong{Monat}  month
\item \strong{Arbeitslose} unemployed
\item \strong{Arbeitslosenquote aller zivilen Erwerbspersonen} unemployment rate based on the total civilian workforce
\item \strong{Arbeitslosenquote d. Abhaengigen ziv. Erwerbspers.} unemployment rate based on dependant civilian working people
\item \strong{Gemeldete Stellen} registered vacancies
\item \strong{Kurzarbeiter} short-time working
\item \strong{Kurzarbeitende Betriebe} companies that implemented short-time work
}
}
\source{
\href{https://www-genesis.destatis.de/genesis/online/}{GENESIS-Online Database}
}
\usage{
destatis13211_0002
}
\description{
Number of males and females by age.
}
\keyword{datasets}
