% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdups.R
\name{dupFiles}
\alias{dupFiles}
\alias{Rdups}
\title{Find duplicate files}
\usage{
dupFiles(files, ...)

Rdups(files, ...)
}
\arguments{
\item{files}{character: file name(s)}

\item{...}{further parameters given to \code{\link[digest:digest]{digest::digest()}}}
}
\value{
a list of file names with the same checksum or \code{NULL}
}
\description{
\code{dupFiles} computes checksums to find duplicate files.
}
\examples{
if (interactive()) {
  files <- list.files(pattern="*.R$", full.names=TRUE, recursive=TRUE)
  dupFiles(files)
}
}
