% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fatalities_statlib}
\alias{fatalities_statlib}
\title{fatalities_statlib}
\format{
A data frame with 40 rows and 3 variables:
\enumerate{
\item \strong{YR} year numbered from 1 to 40
\item \strong{NM} fatalities per 100 000 inhabitants in New Mexico
\item \strong{US} fatalities per 100 000 inhabitants in the US
}
}
\source{
StatLib?
}
\usage{
fatalities_statlib
}
\description{
Fatalities per 100 000 inhabitants in New Mexico and the US about forty years.
}
\keyword{datasets}
