% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghc.R
\name{ghc}
\alias{ghc}
\title{Creates a \code{ghdecompose} pbject}
\usage{
ghc(...)
}
\arguments{
\item{...}{list(s) of filenmaes}
}
\value{
a \code{ghdecompose} pbject
}
\description{
\code{ghc} creates from a list of file names using \code{\link[=ghdecompose]{ghdecompose()}} and deletes mssing files.
}
\examples{
ghc(list.files(system.file(package="mmstat4"), recursive=TRUE))
}
