% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghdecompose.R
\name{ghdecompose}
\alias{ghdecompose}
\title{ghdecompose}
\usage{
ghdecompose(files, dirs = FALSE)
}
\arguments{
\item{files}{character vector: path of files}

\item{dirs}{logical: directory or files names (default: \code{FALSE})}
}
\value{
a data frame with five variables
}
\description{
Decomposes a path of a set of files (or dirs) in several parts:
}
\details{
\itemize{
\item \code{outpath} the path part which is common to all files (basically the place where the ZIP file was extracted)
\item \code{inpath} the path part which is not necessary for a unique address in teh ZIP file
\item \code{minpath} the minimal path part such that all files addressable in unique manner,
\item \code{filename} the basename of the file, and
\item \code{source} the input to \code{shortpath}.
}
}
\examples{
ghget("local")
pdf <- ghdecompose(ghlist(full.names=TRUE))
pdf
}
