% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghfile.R
\name{ghfile}
\alias{ghfile}
\title{ghfile}
\usage{
ghfile(x)
}
\arguments{
\item{x}{character: file names}
}
\value{
the full matching file
}
\description{
Finds either a unique match in the list of files or throws an error with possible candidate files.
}
\examples{
ghfile("data/BANK2.sav")
if (interactive()) ghfile("data/BANK2.SAV")  # throws an error
}
