% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghget.R
\name{ghget}
\alias{ghget}
\title{ghget}
\usage{
ghget(
  dir = getOption("mmstat.dir", tempdir()),
  url = getOption("mmstat.repo",
    "https://github.com/sigbertklinke/mmstat4/archive/refs/heads/main.zip"),
  quiet = !interactive()
)
}
\arguments{
\item{dir}{character: directory name (default: \code{tempdir()})}

\item{url}{character: zip file name or repository, default is the mmstat4 repository}

\item{quiet}{logical: should the download progess be shown (default: \code{!interactive()})}
}
\value{
nothing
}
\description{
Downloads a GitHub repository to a temporary directory, if none given
}
\examples{
if (interactive()) {
  ghget()            # download to tempdir()
  ghget("mytempdir") # download to "mytempdir"
}
}
