% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghlist.R
\name{ghlist}
\alias{ghlist}
\title{ghlist}
\usage{
ghlist(...)
}
\arguments{
\item{...}{parameters given to \code{\link[base:grep]{base::grepl()}}}
}
\value{
character vector of short names
}
\description{
Returns unique (short) names to access each file in the repository.
}
\examples{
if (interactive()) ghlist()
}
