% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghload.R
\name{ghload}
\alias{ghload}
\title{ghload}
\usage{
ghload(x, ...)
}
\arguments{
\item{x}{character(1): name of the file, app or data set}

\item{...}{further parameters used in \code{\link[rio:import]{rio::import()}}}
}
\value{
a data set, open a file in RStudio or runs a shiny app
}
\description{
Loads a data file via \code{\link[rio:import]{rio::import()}} into R from a previously downloaded from a GitHub repository.
}
\examples{
if (interactive()) x <- ghload("bank2.SAV")
}
