% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghopenAddin.R
\name{ghopenAddin}
\alias{ghopenAddin}
\title{ghopenAddin}
\usage{
ghopenAddin()
}
\value{
nothing
}
\description{
A RStudio addin to open a file from the downloaded zip file.
}
\examples{
if (interactive()) ghopenAddin()
}
