% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghrepos.R
\name{ghrepos}
\alias{ghrepos}
\title{ghrepos}
\usage{
ghrepos()
}
\value{
a data frame with the data about the repositories
}
\description{
Retuuns the known repositories and where there are stored. If \code{dir==''} then \code{tempdir()} is used for storage.
}
\examples{
ghrepos()
}
