% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gluehlampen}
\alias{gluehlampen}
\title{gluehlampen}
\format{
A data frame with 4 rows and 6 variables:
\enumerate{
\item \strong{k_u} lower class limit
\item \strong{k_o} upper class limit
\item \strong{h_hat} absolute frequency
\item \strong{f_hat} relative frequency
\item \strong{H_hat} cumulative absolute frequency
\item \strong{F_hat} cumulative relative frequency
}
}
\usage{
gluehlampen
}
\description{
Articial grouped data set about burning time of light bulbs in hours.
}
\keyword{datasets}
