% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haushalte_berlin}
\alias{haushalte_berlin}
\title{haushalte_berlin}
\format{
A data frame with 54 rows and 7 variables:
\enumerate{
\item \strong{Jahr}  year
\item \strong{Einpersonenhaushalte}  number of one person housholds (in 1000)
\item \strong{Zweipersonenhaushalte}  number of two person housholds (in 1000)
\item \strong{Dreipersonenhaushalte}  number of three person housholds (in 1000)
\item \strong{Vierpersonenhaushalte} number of four person housholds (in 1000)
\item \strong{5 und mehr} number of five or more person housholds (in 1000)
\item \strong{Privathaushalte}  total number of households (in 1000)
}
}
\source{
\href{https://www-genesis.destatis.de/genesis/online/}{GENESIS-Online Database} 12211-9034: Privathaushalte: Bundesländer, Jahre (bis 2019), Haushaltsgröße
}
\usage{
haushalte_berlin
}
\description{
Development of the number of households by household size in Berlin.
}
\keyword{datasets}
