% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hhD}
\alias{hhD}
\title{hhD}
\format{
A data frame with 54 rows and 7 variables:
\enumerate{
\item \strong{Jahr} year
\item \strong{Gesamt} total number of households
\item \strong{Einpersonen}  percentage of one person households
\item \strong{Zweipersonen}  percentage of two person households
\item \strong{Dreipersonen}  percentage of three person households
\item \strong{Vierpersonen}  percentage of four person households
\item \strong{Fuenf und mehr Personen}  percentage of five or more person households
}
}
\source{
\href{https://www-genesis.destatis.de/genesis/online/}{GENESIS-Online Database}
}
\usage{
hhD
}
\description{
Development of household shares by household size in Germany.
}
\keyword{datasets}
