% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{install}
\usage{
install(rlibs = NULL, pymods = NULL)
}
\arguments{
\item{rlibs}{character: names of R packages}

\item{pymods}{character: names of Python modules}
}
\value{
invisibly a list of the R packages and Python modules that were attempted to be installed
}
\description{
If a R package or Python module is not already installed, it will be installed with the user's consent.
}
\examples{
if (interactive()) {
  install(rlibs="A3", pymods="numpy")
}
}
