% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kriegsschiffe}
\alias{kriegsschiffe}
\title{kriegsschiffe}
\format{
An object of class \code{data.frame} with 114 rows and 12 columns.
}
\source{
\href{https://de.wikibooks.org/wiki/Benutzer:Philipendula}{User Philipendula from the german wikipedia}
}
\usage{
kriegsschiffe
}
\description{
The dataset describes various characteristics of artillery ships in service at the time of the Second World War.
}
\details{
#' @format A data frame with 114 rows and 12 variables:
\enumerate{
\item \strong{KlasseLand} Ship name and country
\item \strong{Art} Ship type (1=battleship, 3=heavy cruiser, 4=light cruiser, 5=destroyer)
\item \strong{Stapellauf} Ship launch (year)
\item \strong{Verdraengung} Ship displacement (in 1000 tons)
\item \strong{Laenge} Ship length (in m)
\item \strong{Breite} Ship beam (in m)
\item \strong{Tiefgang} Ship draught (in m)
\item \strong{Panzerung} Ship armour
\item \strong{Maschine} Ship engine (in 1000 PS)
\item \strong{Geschwindigkeit} Ship speed (in knots)
\item \strong{Aktionsradius} Ship action radius
\item \strong{Besatzung} Ship crew
}
}
\keyword{datasets}
