% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lotto}
\alias{lotto}
\title{German lotto numbers}
\format{
A data frame with 4181 rows and 10 variables:
\describe{
\item{Datum}{date of drawing}
\item{Wochentag}{week day of drawing (in german)}
\item{Gewinnzahlen}{first drawn number}
\item{V4}{second drawn number}
\item{V5}{third drawn number}
\item{V6}{fourth drawn number}
\item{V7}{fifth drawn number}
\item{V8}{sixth drawn number}
\item{Zusatzzahl}{bonus number, always \code{NA}}
\item{Superzahl}{super number, always \code{NA}}
}

A data frame with 4181 rows and 10 variables:
\describe{
\item{Datum}{date of drawing}
\item{Wochentag}{week day of drawing (in german)}
\item{Gewinnzahlen}{first drawn number}
\item{V4}{second drawn number}
\item{V5}{third drawn number}
\item{V6}{fourth drawn number}
\item{V7}{fifth drawn number}
\item{V8}{sixth drawn number}
\item{Zusatzzahl}{bonus number, always \code{NA}}
\item{Superzahl}{super number, always \code{NA}}
}
}
\usage{
lotto

lotto
}
\description{
Lotto numbers 6 out of 49 including bonus and super number starting from 9 Oct 1955.

Lotto numbers 6 out of 49 including bonus and super number starting from 9 Oct 1955.
}
\keyword{datasets}
