% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lottozahlen}
\alias{lottozahlen}
\title{German lotto numbers
Lotto numbers 6 out of 49 including bonus and super number starting from 9 Oct 1955.}
\format{
A data frame with 4181 rows and 10 variables:
\enumerate{
\item \strong{Datum}  date of drawing
\item \strong{Wochentag}   week day of drawing (in german)
\item \strong{Gewinnzahlen}   first drawn number
\item \strong{V4}   second drawn number
\item \strong{V5}   third drawn number
\item \strong{V6}   fourth drawn number
\item \strong{V7}   fifth drawn number
\item \strong{V8}   sixth drawn number
\item \strong{Zusatzzahl}   bonus number, always \code{NA}
\item \strong{Superzahl}   super number, always \code{NA}
}
}
\usage{
lottozahlen
}
\description{
German lotto numbers
Lotto numbers 6 out of 49 including bonus and super number starting from 9 Oct 1955.
}
\keyword{datasets}
