% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marathon}
\alias{marathon}
\title{marathon}
\format{
A data frame with 28764 rows and 7 variables:
\enumerate{
\item \strong{AGE} age
\item \strong{SEX} Sex (M=male, F=female)
\item \strong{TIME} time in difftime
\item \strong{HOURS} time in hours
\item \strong{AGECAT8} age categories (1=15-24, 2=25-39, 3=40-44, 4=45-49, 5=50-54, 6=55-59, 7=60-64, 8=65+)
\item \strong{AGECAT6} age categories (1=15-44, 2=45-49, 3=50-54, 4=55-59, 5=60-64, 6=65+)
\item \strong{NAME} name or blank
}
}
\source{
\href{http://www.norusis.com/book_DA_v19.php}{Norušis, Marija J. IBM SPSS Statistics 19: Guide to Data Analysis. Internat. ed. Upper Saddle River: Prentice Hall, 2011.}
}
\usage{
marathon
}
\description{
Data about the marathon at 07.10.2001 in Chicago.
}
\keyword{datasets}
