% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normpathes.R
\name{normpathes}
\alias{normpathes}
\title{normpathes}
\usage{
normpathes(x)
}
\arguments{
\item{x}{file pathes}
}
\value{
A list of the same length as \code{x}, the i-th element of which contains the vector of
splits of \code{x[i]}.
}
\description{
Returns a list with normalized pathes.
}
\examples{
normpathes("CRAN/../mmstat4/python/./ghdist.R")
}
