% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{olympic}
\alias{olympic}
\title{olympic}
\format{
A data frame with 33 rows and 12 variables:
\enumerate{
\item \strong{Nr} place in the competition
\item \strong{@100} 100 m run
\item \strong{long} long jump
\item \strong{poid} shotput
\item \strong{haut} high jump
\item \strong{@400} 400 m run
\item \strong{@110} 110-meter hurdles
\item \strong{disq} discus throw
\item \strong{perc} pole vault
\item \strong{jave} javelin
\item \strong{@1500} 1500 m run
\item \strong{score} final points scores
}
}
\source{
\itemize{
\item Example 357 in: Hand, D.J., Daly, F., Lunn, A.D., McConway, K.J. and Ostrowski, E. (1994) A handbook of small data sets, Chapman & Hall, London. 458 p.
\item Lunn, A. D. and McNeil, D.R. (1991) Computer-Interactive Data Analysis, Wiley, New York
}
}
\usage{
olympic
}
\description{
This data set gives the performances of 33 men's decathlon at the Olympic Games (1988).
}
\seealso{
\link[ade4:olympic]{ade4::olympic}
}
\keyword{datasets}
