% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rlibs.R
\name{packages}
\alias{packages}
\alias{modules}
\alias{Rlibs}
\title{Extract \code{library} and \code{require} calls in R and \code{import} calls from Python}
\usage{
packages(files)

Rlibs(files)

modules(files)
}
\arguments{
\item{files}{character: file name(s)}
}
\value{
a table how frequently the packages are called
}
\description{
\code{packages} counts the number of \code{library}/\code{require}/\code{import} calls for R and Python commands in the files.
}
\examples{
if (interactive()) {
  files <- list.files(pattern="*.(R|py)$", full.names=TRUE, recursive=TRUE)
  packages(files)
}
}
