% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pechstein}
\alias{pechstein}
\title{pechstein}
\format{
A data frame with 29 rows and 3 variables:
\enumerate{
\item \strong{Datum} Date of control
\item \strong{Tag} Date of control as days from 1st Jan 2000
\item \strong{Retikulozyten} Reticulocytes levels
}
}
\source{
Bild.de, 17.07.2009, behind paywall: https://www.bild.de/sport/mehr-sport/claudia-pechstein/was-verraten-ihre-blutwerte-im-doping-skandal-8932510.bild.html
}
\usage{
pechstein
}
\description{
Claudia Pechstein is a German speed skater which was banned because of blood doping.
The ban was based on irregular levels of reticulocytes in her blood. The data set is a
non-random excerpt of 95 blood controls since 2000.
}
\keyword{datasets}
