% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pechstein_isu}
\alias{pechstein_isu}
\title{pechstein_isu}
\format{
A data frame with 29 rows and 3 variables:
\enumerate{
\item \strong{no} Number of control from 1 to 95
\item \strong{date} Date of control as \code{Date}
\item \strong{place} Place of control, \verb{(Out of Competition)} means a control during training
\item \strong{reti} Reticulocytes levels in percent
}
}
\source{
\href{https://www.jensweinreich.de/2009/07/08/die-blutkontrollen-von-claudia-pechstein/}{Blog of Jens Weinreich, 08.07.2009}
}
\usage{
pechstein_isu
}
\description{
Claudia Pechstein is a German speed skater which was banned because of blood doping.
The ban was based on irregular levels of reticulocytes in her blood. The data set
with the full values from the 95 blood controls since 2000. The threshold value
the ISU used was 2.4.
}
\keyword{datasets}
