% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgMissing.R
\name{pkgMissing}
\alias{pkgMissing}
\title{pkgMissing}
\usage{
pkgMissing(package)
}
\arguments{
\item{package}{character: string naming the package/name space to load.}
}
\value{
a logical value
}
\description{
Checks if a \code{package} is available.
}
\examples{
pkgMissing("tools")
pkgMissing("A3")
}
