% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prg.R
\name{prg}
\alias{prg}
\title{prg}
\usage{
prg(file = NULL, pattern = NULL, n = 5, ...)
}
\arguments{
\item{file}{character: a file name (default: \code{NULL})}

\item{pattern}{character: regular expression to be matched in the file names (default: \code{NULL})}

\item{n}{integer: number of similar files shown if no \code{file} can be found (default: \code{5})}

\item{...}{further parameters given to \code{grepl}}
}
\value{
the complete path including the file name(s)
}
\description{
Returns a list of matching names of an example file or app in \code{mmstat4}. Note that \code{file}
is interpreted as a regular expression.
}
\examples{
prg() # all available example files
prg("mmstat/lottozahlen.R")
if(interactive()) {
  edit(prg("mmstat/lottozahlen.R"))       # open text editor
  file.edit(prg("mmstat/lottozahlen.R"))  # open in RStudio
}
}
