% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rlibs.R
\name{py_env}
\alias{py_env}
\title{py_env}
\usage{
py_env()
}
\value{
the name of the virtual Python environment currently used by \code{mmstat4}
}
\description{
Name of the currently used virtual emvironment.
}
\examples{
py_env()
}
