% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsolo.R
\name{Rsolo}
\alias{Rsolo}
\title{Rsolo}
\usage{
Rsolo(files, start = 1, path = NULL)
}
\arguments{
\item{files}{character: file name(s)}

\item{start}{integer: numeric index from which file to start (default: \code{1})}

\item{path}{character: path to start from (default: \code{getwd()})}
}
\value{
nothing
}
\description{
Assume that all files given are valid R files which should run indepedently
of each other. If an error appears the execution is stopped.
}
\examples{
if (interactive()) {
  files <- list.files(patter="*.R$", full.names=TRUE, recursive=TRUE)
  Rsolo(files)
}
}
