% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{run}
\usage{
run(name, ...)
}
\arguments{
\item{name}{character: name of example file}

\item{...}{further parameters to \code{source}}
}
\value{
nothing
}
\description{
\code{source}s the example file
}
\examples{
run("mmstat/lottozahlen.R")
}
