% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{salary}
\alias{salary}
\title{salary}
\format{
A data frame with 474 rows and 11 variables:
\enumerate{
\item \strong{id} Ssquence number
\item \strong{salbeg}  beginning salary (US$)
\item \strong{sex} sex of employee (males, females)
\item \strong{time} job seniority (years)
\item \strong{age} age of employee (years)
\item \strong{salnow} current salary (US$)
\item \strong{edlevel} educational level (years)
\item \strong{work}  work experience (years)
\item \strong{jobcat}  employment category (clerical/office trainee/security officer/college trainee/exempt employee/mba trainee/technical)
\item \strong{minority} minority classification (white/nonwhite)
\item \strong{sexrace} sex & race classification (white males/minority males/white females/minority females)
}
}
\usage{
salary
}
\description{
The data set contains information about 474 employees hired by the Harris Trust and Savings Bank between 1969 and 1971. The bank was involved into a Equal Employment Opportunity litigation were the data was used.
}
\references{
\itemize{
\item Roberts, H.V. (1979) \emph{An analysis of employee compensation.} Report 7946, Center for Mathematical Studies in Business and Economics, University of Chicago.
\item Roberts, H.V. (1980) \emph{Statistical bases in the measurement of employment discrimination.} In: Comparable Worth: Issues and Alternatives by E.R. Livernash (ed.) Washington, D.C.: Equal Employment Advisory Council.
}
}
\keyword{datasets}
