% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sas_body}
\alias{sas_body}
\title{sas_body}
\format{
A data frame with 237 rows and 4 variables:
\enumerate{
\item \strong{Sex} Gender
\item \strong{Age} Age in month
\item \strong{Height} Height in inches
\item \strong{Weight} Weight in pounds
}
}
\source{
SAS online documentation V8
}
\usage{
sas_body
}
\description{
The data set consists of data about children taken from the SAS online documentation V8:
}
\keyword{datasets}
