% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show,character-method}
\alias{show,character-method}
\title{show}
\usage{
\S4method{show}{character}(object)
}
\arguments{
\item{object}{character: name of example program}
}
\value{
nothing
}
\description{
Edits a program from the package \code{mmstat4} using \code{file.edit}.
Note that the file is copied to a temporary directory, such that you can not overwrite the original file!
If you do not save modified code at another location then the next \code{show} will overwrite your modified code.
}
\examples{
if (interactive()) {
  show("stat/sum.R")
}
}
