% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{staedtemietenr}
\alias{staedtemietenr}
\title{staedtemietenr}
\format{
A data frame with 16777 rows and 7 variables:
\enumerate{
\item \strong{Thema} Group
\item \strong{Stadt} City
\item \strong{Stadtbezirk} Muncipality
\item \strong{Zimmer} Number of rooms
\item \strong{Etage} Floor
\item \strong{Miete} Basic rent in EUR
\item \strong{Flaeche} Living space in m2
}
}
\usage{
staedtemietenr
}
\description{
Housing offers in almost forty German cities.
}
\keyword{datasets}
