% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thousand_songs}
\alias{thousand_songs}
\title{thousand_songs}
\format{
A data frame with 994 rows and 5 variables:
\enumerate{
\item \strong{THEME} Song theme (Heartbreak/Life and death/Love/Party songs/People and places/Politics and protest/Sex)
\item \strong{TITLE} Song title
\item \strong{ARTIST} Song artist
\item \strong{YEAR} Year of publishing
\item \strong{SPOTIFY_URL} Spotify URL or blank
}
}
\source{
\href{https://www.kaggle.com/iampedroalz/1000songs}{Kaggle: 1000 songs everyone must hear - The Guardian's songs list}
}
\usage{
thousand_songs
}
\description{
Selection of the 1000 songs everyone must hear performed by The Guardian in the year 2009.
}
\keyword{datasets}
