% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic3}
\alias{titanic3}
\title{titanic3}
\format{
A data frame with 1309 rows and 14 variables:
\enumerate{
\item \strong{pclass} a factor with levels 1st, 2nd, and 3rd
\item \strong{survived} survival with 0 = No and  1 = Yes
\item \strong{name} name of passenger
\item \strong{sex} a factor with levels female and male
\item \strong{age} age in years
\item \strong{sibsp} number of Siblings/Spouses Aboard
\item \strong{parch} number of Parents/Children Aboard
\item \strong{ticket} ticket number
\item \strong{fare} passenger fare
\item \strong{cabin} cabin
\item \strong{embarked} a factor with levels Cherbourg, Queenstown, and Southampton
\item \strong{boat} lifeboat number
\item \strong{body} body identification number
\item \strong{home.dest} home/destination
}
}
\source{
http://biostat.mc.vanderbilt.edu/twiki/pub/Main/DataSets/titanic.html
}
\usage{
titanic3
}
\description{
The sinking of the RMS Titanic is one of the most infamous shipwrecks in history. On April 15, 1912, during her maiden voyage, the Titanic sank after colliding with an iceberg, killing 1502 out of 2224 passengers and crew.
The data set describes the survival status of individual passengers on the Titanic. The data set does not contain information for the crew, but it does contain actual and estimated ages for almost 80\% of the passengers.
}
\references{
\itemize{
\item Harrell, F. E. 2001. Regression Modeling Strategies with Applications to Linear Models, Logistic Regression, and Survival Analysis. Springer.
\item Ugarte, M. D., Militino, A. F., and Arnholt, A. T. 2015. Probability and Statistics with R, Second Edition. Chapman & Hall / CRC.
}
}
\seealso{
\link[PASWR2:TITANIC3]{PASWR2::TITANIC3}
}
\keyword{datasets}
