% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic_full}
\alias{titanic_full}
\title{titanic_full}
\format{
A data frame with 2201 rows and 4 variables:
\enumerate{
\item \strong{Class} 1=1st, 2=2nd, 3=3rd, 4=Crew
\item \strong{Sex} 1=Male, 2=Female
\item \strong{Age} 1=Child, 2=Adult
\item \strong{Survived} 1=No, 2=Yes
}
}
\source{
Dawson, Robert J. MacG. (1995), The ‘Unusual Episode’ Data Revisited. Journal of Statistics Education, 3. doi: \href{https://www.tandfonline.com/doi/full/10.1080/10691898.1995.11910499}{10.1080/10691898.1995.11910499}.
}
\usage{
titanic_full
}
\description{
This data set provides information on the fate of passengers on the fatal maiden voyage of the ocean liner ‘Titanic’, according to economic status (class), sex, age and survival.
}
\seealso{
\link[datasets:Titanic]{datasets::Titanic}
}
\keyword{datasets}
