% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlExists.R
\name{urlExists}
\alias{urlExists}
\title{urlExists}
\usage{
urlExists(url)
}
\arguments{
\item{url}{a vector of text URLs}
}
\value{
\code{TRUE} if URL exists otherwise \code{FALSE}
}
\description{
Verifies whether a provided \code{url} is downloadable, without detecting redirections in the URL.
}
\examples{
if (interactive()) {
  urlExists("https://hu-berlin.de/sk")
  urlExists("https://huglawurza.de")
}
}
