% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghget.R
\name{ghget}
\alias{ghget}
\alias{ghset}
\title{ghget}
\usage{
ghget(key = "mmstat4", force = FALSE)

ghset(key, url, install = TRUE)
}
\arguments{
\item{key}{character: directory name (default: \code{mmstat4})}

\item{force}{logical: download and unzip in any case? (default: \code{FALSE})}

\item{url}{character: URL to zip file of repository, default is the mmstat4 repository}

\item{install}{logical: call \code{ghget} in \code{ghset}? (default: \code{TRUE})}
}
\value{
nothing
}
\description{
\itemize{
\item \code{ghset} set for a key the installation directory and GitHubs Zip URL. If \code{dir} is \code{NULL} then \code{tempdir()} is used.
\item \code{ghget} downloads the GitHub repository and unzips it if necessary.
}
}
\examples{
if (interactive()) {
  ghget("dummy")
  ghget()        # or ghget("mmstat4")
}
}
