% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghget.R
\name{ghget}
\alias{ghget}
\title{ghget}
\usage{
ghget(..., .force = FALSE)
}
\arguments{
\item{...}{parameters to set and activate a repository}

\item{.force}{logical: download and unzip in any case? (default: \code{FALSE})}
}
\value{
the name of the activated directory
}
\description{
Makes a repository the active repository and downloads it if neceassary.
If the function is run interactively then you are asked if you want to store the repository in the application directory
\code{\link[rappdirs:user_data_dir]{rappdirs::user_data_dir()}} for \code{mmstat4} or in the temporary directory \code{\link[=tempdir]{tempdir()}}. Otherwise the stored directory
path is used (usually \code{\link[=tempdir]{tempdir()}}).
}
\details{
Note, the list of repository names, directories and urls is stored in the installation directory, too.
}
\examples{
if (interactive()) {

}
}
