% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghopen.R
\name{ghopen}
\alias{ghopen}
\title{ghopen}
\usage{
ghopen(x, ...)
}
\arguments{
\item{x}{character(1): name of the file, app or data set}

\item{...}{further parameters}
}
\value{
a data set, open a file in RStudio or runs a shiny app
}
\description{
Opens a R or Rmarkdown into the editor, starts a Shiny app or opens a data set
previously downloaded from a GitHub repository.
}
\examples{
if (interactive()) x <- ghopen("bank2.SAV")
}
