% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghpath.R
\name{ghpath}
\alias{ghpath}
\title{ghpath}
\usage{
ghpath(df, from = c("minpath", "commonpath", "uniquepath", "filename"))
}
\arguments{
\item{df}{data frame: returned from \code{ghdecompose}}

\item{from}{character: either \code{minpath} (defaulkt), \code{commonpath}, \code{uniquepath}, or \code{filename}}
}
\value{
a character vector with file pathes
}
\description{
Returns a path for files based on \code{ghdecompose}.
}
\examples{
ghget("dummy")
pdf <- ghdecompose(ghlist(full.names=TRUE))
ghpath(pdf)
ghpath(pdf, 'c') # equals the input to ghdecompose
ghpath(pdf, 'u')
ghpath(pdf, 'm')
ghpath(pdf, 'f')
}
