% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghquery.R
\name{ghquery}
\alias{ghquery}
\title{gquery}
\usage{
ghquery(query, n = 6, full.names = FALSE)
}
\arguments{
\item{query}{character: query string}

\item{n}{integer: maximal number of matches to return}

\item{full.names}{logical: should full names used instead of short names (default: \code{FALSE})}
}
\value{
character vector of short names fitting best to the query
}
\description{
Queries the unique (short) names for each file in the repository.
}
\examples{
if (interactive()) ghquery("bank")
}
