% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghrepos.R
\name{ghrepos}
\alias{ghrepos}
\title{ghrepos}
\usage{
ghrepos(key = NULL)
}
\arguments{
\item{key}{character: "name" of the repository to find (default: \code{NULL})}
}
\value{
a data frame with the data about the repositories
}
\description{
If \code{key} is \code{NULL}, then it returns the known repositories and where they are stored.
If \code{key} is not \code{NULL}, then possible addresses for a repository are returned .
}
\examples{
ghrepos()
}
