% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prg.R
\name{prg}
\alias{prg}
\title{prg}
\usage{
prg(file = NULL)
}
\arguments{
\item{file}{character: file name (defauilt. \code{NULL})}
}
\value{
the complete path including the file name
}
\description{
Returns the full name of an example file in \code{mmstat4} or a list of available example files.
}
\examples{
prg() # all available example files
prg("mmstat/lottozahlen.R")
if(interactive()) {
  edit(prg("mmstat/lottozahlen.R"))        # open text editor
  file.edit (prg("mmstat/lottozahlen.R"))  # open in RStudio
}
}
