% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{run}
\usage{
run(name, ...)
}
\arguments{
\item{name}{character: name of example file}

\item{...}{further parameters to \code{\link[base:source]{base::source()}} or \link[shiny:runApp]{shiny::runApp}}
}
\value{
nothing
}
\description{
\code{source}s the example file. The parameter \code{echo} from \code{source} is set to \code{TRUE} by default.
}
\examples{
run("mmstat/lottozahlen.R")
run("mmstat/lottozahlen.R", echo=FALSE)
}
