% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show,character-method}
\alias{show,character-method}
\title{show}
\usage{
\S4method{show}{character}(object)
}
\arguments{
\item{object}{character: name of example program}
}
\value{
nothing
}
\description{
Edits a program from the package \code{mmstat} using \code{file.edit}.
}
\examples{
if (interactive()) {
  show("stat/sum.R")
}
}
